Changed defaultHeight for Staves
I changed the page height for the staves and a lot of staves and notes were being cut off from the bottom of the page, and were not visible.

I made the height the same in all of the javascript files to keep all the pages in a standard format.